+++ src/lib/dfp/prepare-a9.ts
@@ -10,14 +10,13 @@ import { isGoogleProxy } from 'lib/detect/detect-google-proxy';
 import { isInCanada } from 'lib/utils/geo-utils';
 import { a9 } from '../header-bidding/a9/a9';
 import { shouldIncludeOnlyA9 } from '../header-bidding/utils';
-import { dfpEnv } from './dfp-env';
 
 const shouldLoadA9 = () =>
 	// There are two articles that InfoSec would like to avoid loading scripts on
 	!commercialFeatures.isSecureContact &&
 	!isGoogleProxy() &&
-	dfpEnv.hbImpl.a9 &&
-	commercialFeatures.dfpAdvertising &&
+	window.guardian.config.switches.a9HeaderBidding &&
+	commercialFeatures.shouldLoadGoogletag &&
 	!commercialFeatures.adFree &&
 	!window.guardian.config.page.hasPageSkin &&
 	!isInCanada();