+++ src/lib/dfp/prepare-prebid.spec.ts
@@ -8,7 +8,6 @@ import { log } from '@guardian/libs';
 import { commercialFeatures } from 'lib/commercial-features';
 import { isInCanada } from 'lib/utils/geo-utils';
 import { prebid } from '../header-bidding/prebid/prebid';
-import { dfpEnv } from './dfp-env';
 import { _ } from './prepare-prebid';
 
 const { setupPrebid } = _;
@@ -133,13 +132,15 @@ describe('init', () => {
 	beforeEach(() => {
 		jest.resetAllMocks();
 		fakeUserAgent();
+		window.guardian.config.switches = {};
 	});
 
 	it('should initialise Prebid when Prebid switch is ON and advertising is on and ad-free is off', async () => {
 		expect.hasAssertions();
-
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(tcfv2WithConsent);
 		mockGetConsentFor(true);
@@ -150,9 +151,10 @@ describe('init', () => {
 
 	it('should not initialise Prebid when useragent is Google Web Preview', async () => {
 		expect.hasAssertions();
-
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		fakeUserAgent('Google Web Preview');
 		mockOnConsent(tcfv2WithConsent);
@@ -165,9 +167,11 @@ describe('init', () => {
 	it('should not initialise Prebid when no header bidding switches are on', async () => {
 		expect.hasAssertions();
 
-		commercialFeatures.dfpAdvertising = true;
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
-		dfpEnv.hbImpl = { prebid: false, a9: false };
+		window.guardian.config.switches = {
+			prebidHeaderBidding: false,
+		};
 		mockOnConsent(tcfv2WithConsent);
 		mockGetConsentFor(true);
 
@@ -178,8 +182,10 @@ describe('init', () => {
 	it('should initialise Prebid when NOT in Canada', async () => {
 		expect.hasAssertions();
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(tcfv2WithConsent);
 		mockGetConsentFor(true);
@@ -192,8 +198,10 @@ describe('init', () => {
 	it('should NOT initialise Prebid when in Canada', async () => {
 		expect.hasAssertions();
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(tcfv2WithConsent);
 		mockGetConsentFor(true);
@@ -206,8 +214,10 @@ describe('init', () => {
 	it('should not initialise Prebid when advertising is switched off', async () => {
 		expect.hasAssertions();
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = false;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = false;
 		commercialFeatures.adFree = false;
 		mockOnConsent(tcfv2WithConsent);
 		mockGetConsentFor(true);
@@ -219,8 +229,10 @@ describe('init', () => {
 	it('should not initialise Prebid when ad-free is on', async () => {
 		expect.hasAssertions();
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = true;
 		mockOnConsent(tcfv2WithConsent);
 		mockGetConsentFor(true);
@@ -232,8 +244,10 @@ describe('init', () => {
 	it('should not initialise Prebid when the page has a pageskin', async () => {
 		expect.hasAssertions();
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		window.guardian.config.page.hasPageSkin = true;
 		mockOnConsent(tcfv2WithConsent);
@@ -244,8 +258,10 @@ describe('init', () => {
 	});
 
 	it('should initialise Prebid when the page has no pageskin', async () => {
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		window.guardian.config.page.hasPageSkin = false;
 		mockOnConsent(tcfv2WithConsent);
@@ -258,8 +274,10 @@ describe('init', () => {
 	it('should initialise Prebid if TCFv2 consent with correct Sourcepoint Id is true ', async () => {
 		expect.hasAssertions();
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(tcfv2WithConsent);
 		mockGetConsentFor(true);
@@ -270,8 +288,10 @@ describe('init', () => {
 	it('should not initialise Prebid if TCFv2 consent with correct Sourcepoint Id is false', async () => {
 		expect.assertions(2);
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(tcfv2WithoutConsent);
 		mockGetConsentFor(false);
@@ -289,8 +309,10 @@ describe('init', () => {
 	it('should initialise Prebid in CCPA if doNotSell is false', async () => {
 		expect.hasAssertions();
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(ccpaWithConsent);
 		mockGetConsentFor(true);
@@ -301,8 +323,10 @@ describe('init', () => {
 	it('should not initialise Prebid in CCPA if doNotSell is true', async () => {
 		expect.assertions(2);
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(ccpaWithoutConsent);
 		mockGetConsentFor(false);
@@ -320,8 +344,10 @@ describe('init', () => {
 	it('should initialise Prebid in AUS if Advertising is not rejected', async () => {
 		expect.hasAssertions();
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(ausWithConsent);
 		mockGetConsentFor(true);
@@ -332,8 +358,10 @@ describe('init', () => {
 	it('should not initialise Prebid in AUS if Advertising is rejected', async () => {
 		expect.assertions(2);
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(ausWithoutConsent);
 		mockGetConsentFor(false);
@@ -351,8 +379,10 @@ describe('init', () => {
 	it('should not initialise Prebid if the framework is invalid', async () => {
 		expect.assertions(2);
 
-		dfpEnv.hbImpl = { prebid: true, a9: false };
-		commercialFeatures.dfpAdvertising = true;
+		window.guardian.config.switches = {
+			prebidHeaderBidding: true,
+		};
+		commercialFeatures.shouldLoadGoogletag = true;
 		commercialFeatures.adFree = false;
 		mockOnConsent(invalidWithoutConsent);
 		mockGetConsentFor(true);