+++ src/lib/dfp/prepare-prebid.ts
@@ -10,12 +10,11 @@ import { isGoogleProxy } from 'lib/detect/detect-google-proxy';
 import { isInCanada } from 'lib/utils/geo-utils';
 import { prebid } from '../header-bidding/prebid/prebid';
 import { shouldIncludeOnlyA9 } from '../header-bidding/utils';
-import { dfpEnv } from './dfp-env';
 
 const shouldLoadPrebid = () =>
 	!isGoogleProxy() &&
-	dfpEnv.hbImpl.prebid &&
-	commercialFeatures.dfpAdvertising &&
+	window.guardian.config.switches.prebidHeaderBidding &&
+	commercialFeatures.shouldLoadGoogletag &&
 	!commercialFeatures.adFree &&
 	!window.guardian.config.page.hasPageSkin &&
 	!shouldIncludeOnlyA9 &&